Move getAll to utilities
