+++ src/services/domains/domains.ts
@@ -3,7 +3,6 @@ import { API_ROOT } from 'src/constants';
 import Request,
 {
   CancellableRequest,
-  getAll,
   setData,
   setMethod,
   setParams,
@@ -25,10 +24,7 @@ export const getDomains = (
   setParams(params),
   setXFilter(filters),
 )
-  .then(response => response.data);;
-
-export const getAllDomains = (params?: any, filter?: any): Promise<Domain[]> =>
-  getAll(getDomains, params, filter)
+  .then(response => response.data);
 
 export const getDomains$ = (
   params: any = {},
