+++ src/services/domains/index.ts
@@ -1,7 +1,6 @@
 export {
   getDomains,
   getDomains$,
-  getAllDomains,
   getDomain,
   createDomain,
   updateDomain,
