+++ src/services/images/index.ts
@@ -1,6 +1,5 @@
 export {
   getImages,
-  getAllImages,
   getImage,
   getImagesPage,
   getLinodeImages,
