+++ src/services/linodes/getLinodeInfo.ts
@@ -1,6 +1,3 @@
-import * as Bluebird from 'bluebird';
-import { range } from 'ramda';
-
 import { API_ROOT } from 'src/constants';
 
 import Request, { setMethod, setParams, setURL, setXFilter } from '../index';
@@ -27,28 +24,6 @@ export const getLinodeKernels = (params?: any, filter?: any) =>
     setXFilter(filter)
   )
     .then(response => response.data);
-
-export const getAllKernels: (params?: any, filters?: any) => Promise<Linode.Kernel[]> =
-  (params = {}, filters = {}) => {
-    const pagination = { ...params, page_size: 100 };
-
-    return getLinodeKernels(params, filters)
-      .then(({ data: firstPageData, page, pages }) => {
-
-        // If we only have one page, return it.
-        if (page === pages) { return firstPageData; }
-
-        // Create an iterable list of the remaining pages.
-        const remainingPages = range(page + 1, pages + 1);
-
-        //
-        return Bluebird
-          .map(remainingPages, nextPage =>
-            getLinodeKernels({ ...pagination, page: nextPage }, filters).then(response => response.data),
-          )
-          .then(allPages => allPages.reduce((result, nextPage) => [...result, ...nextPage], firstPageData));
-      });
-  }
 
 export const getLinodeTypes = () =>
   Request<Page<Type>>(
