+++ src/services/nodebalancers/index.ts
@@ -1,6 +1,5 @@
 export {
   getNodeBalancers,
-  getAllNodeBalancers,
   getNodeBalancer,
   getNodeBalancerStats,
   createNodeBalancer,
