+++ src/services/nodebalancers/nodebalancers.ts
@@ -1,6 +1,6 @@
 import { API_ROOT } from 'src/constants';
 
-import Request, { getAll, setData, setMethod, setParams, setURL, setXFilter } from '../index';
+import Request, { setData, setMethod, setParams, setURL, setXFilter } from '../index';
 
 import {
   NodeBalancerSchema,
@@ -19,9 +19,6 @@ export const getNodeBalancers = (pagination: any = {}, filters: any = {}) =>
     setXFilter(filters),
   )
     .then(response => response.data);
-
-export const getAllNodeBalancers = (params?: any, filters?: any): Promise<NodeBalancer[]> =>
-  getAll(getNodeBalancers, params, filters)
 
 export const getNodeBalancer = (id: number) => Request<NodeBalancer>(
   setURL(`${API_ROOT}/nodebalancers/${id}`),
