+++ src/services/volumes/index.ts
@@ -1,7 +1,6 @@
 export {
   getVolume,
   getVolumes,
-  getAllVolumes,
   createVolume,
   updateVolume,
   deleteVolume,
